requirements.txt
django
google-api-python-client 
google-auth 
google-auth-oauthlib 
google-auth-httplib2
pillow
pandas
xlsxwriter
tensorflow
torch 
torchvision 
torchmetrics 
torchaudio 
clip 
albumentations 
scikit-learn 
joblib 
pillow 
numpy 
open_clip_torchd